+++ src/core/settings.ts
@@ -41,14 +41,6 @@ export async function setupSettings(){
 			public: true,
 			section: 'section',
 		},
-		"moveDueDatesToEnd": {
-			label: "Sort tasks without due dates at the end of the list",
-			value: false,
-			type: SettingItemType.Bool,
-			public: true,
-			section: 'section',
-		}
-
 	})
 	await joplin.settings.onChange(setupTimer)
 }