+++ src/core/timer.ts
@@ -12,13 +12,17 @@ import { refreshNoteData } from "./markdown";
 
 /** Variable Initialization ************************************************************************************************************************/
 var timer = null
+var updating = false;
 
 /** refreshInterfaces ********************************************************************************************************************************
  * Refreshes the panel and the notes associated with the current profile                                                                              *
  ***************************************************************************************************************************************************/
  export async function refreshInterfaces(){
+    if (updating) return;
+    updating = true;
     await refreshPanelData()
     await refreshNoteData()
+    updating = false;
 }
 
 export async function setupTimer(){